Fix discord moving
