+++ docs/conf.py
@@ -242,8 +242,6 @@ exclude_patterns = [
     '_api/airflow/providers/cncf/kubernetes/index.rst',
     '_api/airflow/providers/databricks/index.rst',
     '_api/airflow/providers/datadog/index.rst',
-    '_api/airflow/providers/ddiscord/index.rst',
-    '_api/airflow/providers/dindding/index.rst',
     '_api/airflow/providers/dingding/index.rst',
     '_api/airflow/providers/discord/index.rst',
     '_api/airflow/providers/docker/index.rst',
