+++ tests/providers/discord/hooks/__init__.py
